+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 01. Added dev.orne.config.AbstractConfig
 01. Added dev.orne.config.HierarchicalConfig
 01. Added dev.orne.config.AbstractStringConfig
+01. Added dev.orne.config.SystemConfig